// sdk/src/generated/accounts/ammConfig.ts
export type AmmConfig = {
  discriminator: ReadonlyUint8Array;
  /** Bump to identify PDA */
  bump: number;
  index: number;
  /** Address of the protocol owner */
  owner: Address;
  /** The protocol fee */
  protocolFeeRate: number;
  /** The trade fee, denominated in hundredths of a bip (10^-6) */
  tradeFeeRate: number;
  /** The tick spacing */
  tickSpacing: number;
  /** The fund fee, denominated in hundredths of a bip (10^-6) */
  fundFeeRate: number;
  paddingU32: number;
  fundOwner: Address;
  padding: Array<bigint>;
};